visao-computacional/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
matplotlib>=3.7.0
pyyaml>=6.0
supabase>=2.0.0
python-dotenv>=1.0.0
